requirements.txt
numpy
opencv-python
Pillow
pandas
easyocr